Use Node's 'readable' event for processing chunks
Rather than manually parsing each chunk looking for line breaks, Node
provides a 'readable' event that is emitted when a chunk of data is
available for reading. Since each chunk represents a tweet, no
additional processing needs to be performed.

This also consolidates the binding of request events into the Streamer
instance itself.