Catch Twitter Api error messages
The Twitter Streaming API does not return JSON formatted error messaging
when clients have exceeded the rate limit, etc. This causes the stream
parser to throw a JSON parsing error which is emitted as an "error"
event.

Because node-twitter obfuscates the original Twitter error message, it
is difficult for end-users to know if the Twitter Streaming Api is
trying to tell them something.

This detects if the parsing error was due to a malformed JSON payload or
if it was due to a legitimate Api error.